Updated inertia_tensor property + _compute and diagonalize methods
Compute_inertia_tensor needed to be updated because the previous algorithm does not work if the tetrahedral decomposition of the shape includes tetrahedra with negative volumes. This new method, implemented based on an algorithm published by A. M. Messner in 1980 is significantly faster and provides slightly more accurate results. Because of this, the relevant pytest has been updated to better check the method.