Added more stringent pytests for centroid and centroid setter
The centroid methods are core functionality to this class, and prior to this commit was not being tested. This fixes that, with a two-part test that checks for initial and repeated setting behavior. The monte carlo centroid compute has also been updated to test more samples on CircleCI. Test sampling has been tuned to account for the tighter tolerances